fix: dep update

+++ pkg/mapper/services/mapper.go
@@ -16,7 +16,7 @@ func MapCRDConditionsToAPI(conditions []metav1.Condition, currentTime time.Time)
 			Type_:  string(condition.Type),
 			Status: (*testtriggersv1.TestTriggerConditionStatuses)(&condition.Status),
 			Reason: condition.Reason,
-			Ttl:    currentTime.Sub(condition.LastTransitionTime.Time),
+			Ttl:    int32(currentTime.Sub(condition.LastTransitionTime.Time)),
 		})
 	}
